+++ NEWS.md
@@ -1,10 +1,11 @@
-# tern 0.9.10.9007
+# tern 0.9.10.9008
 
 ### Miscellaneous
 * Updated `roxygen2` to 8.0.0 and added `@exportS3Method` tags for S3 methods in `decorate_grob.R` and `utils_grid.R`.
 
 ### Enhancements
-
+* Added `alternative` argument to `s_coxph_pairwise()` to allow one-sided hypothesis testing. 
+* Added `lr_stat_df` to the parameters return list of `s_coxph_pairwise()`.
 * Added `uncond_exact_diff` method to `estimate_proportion_diff()` for the unconditional exact confidence interval for the difference in proportions by inverting one-sided tail tests over a nuisance parameter.
 
 # tern 0.9.10